+++ cx-portal/src/components/pages/Admin/components/RegistrationRequests/index.tsx
@@ -1,48 +1,45 @@
-import React, { useEffect, useState } from 'react'
+import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import {
-  Table,
   Typography,
   PageHeader,
-  Button,
+  PageSnackbar,
 } from 'cx-portal-shared-components'
-import { adminRegistrationSelector } from 'features/admin/registration/slice'
-import { useSelector, useDispatch } from 'react-redux'
-import {
-  fetchRegistrationRequests,
-  fetchCompanyDetail,
-  approveRequest,
-  declineRequest,
-} from 'features/admin/registration/actions'
-import { RegistrationRequestsTableColumns } from 'components/pages/Admin/components/RegistrationRequests/registrationTableColumns'
+import { useDispatch } from 'react-redux'
+import { fetchCompanyDetail } from 'features/admin/registration/actions'
 import './RegistrationRequests.scss'
 import { GridCellParams } from '@mui/x-data-grid'
 import CompanyDetailOverlay from './CompanyDetailOverlay'
 import ConfirmationOverlay from './ConfirmationOverlay/ConfirmationOverlay'
-import uniqueId from 'lodash/uniqueId'
+import {
+  useApproveRequestMutation,
+  useDeclineRequestMutation,
+  useFetchCompanySearchQuery,
+} from 'features/admin/applicationRequestApiSlice'
+import { RequestList } from './components/RequestList'
 
 export default function RegistrationRequests() {
   const { t } = useTranslation()
+
   const dispatch = useDispatch()
 
   const [overlayOpen, setOverlayOpen] = useState<boolean>(false)
-
-  const [currentPage, setCurrentPage] = useState<number>(0)
-  const [pageSize] = useState<number>(10)
 
   const [confirmModalOpen, setConfirmModalOpen] = useState<boolean>(false)
 
   const [selectedRequestId, setSelectedRequestId] = useState<string>()
   const [actionType, setActionType] = useState<string>('approve')
 
-  const { loading, registrationRequests, paginationData } = useSelector(
-    adminRegistrationSelector
-  )
+  const [expr, setExpr] = useState<string>('')
 
-  useEffect(() => {
-    const params = { size: pageSize, page: currentPage }
-    dispatch(fetchRegistrationRequests({ params }))
-  }, [dispatch, currentPage, pageSize])
+  const [approveRequest] = useApproveRequestMutation()
+  const [declineRequest] = useDeclineRequestMutation()
+
+  const [isLoading, setIsLoading] = useState<boolean>(false)
+
+  const [showErrorAlert, setShowErrorAlert] = useState<string>('')
+
+  const [loaded, setLoaded] = useState<number>(0)
 
   const onTableCellClick = (params: GridCellParams) => {
     // Show overlay only when detail field clicked
@@ -63,26 +60,40 @@ export default function RegistrationRequests() {
     setSelectedRequestId(id)
   }
 
-  const makeActionSelectedRequest = async () => {
-    if (actionType === 'approve' && selectedRequestId)
-      await dispatch(approveRequest(selectedRequestId))
-    if (actionType === 'decline' && selectedRequestId)
-      await dispatch(declineRequest(selectedRequestId))
-
-    const params = { size: pageSize, page: 0 }
-    dispatch(fetchRegistrationRequests({ params }))
-
-    setConfirmModalOpen(false)
+  const onErrorAlertClose = () => {
+    setShowErrorAlert('')
   }
 
-  const columns = RegistrationRequestsTableColumns(
-    useTranslation,
-    onApproveClick,
-    onDeclineClick
-  )
+  const makeActionSelectedRequest = async () => {
+    setIsLoading(true)
+    setConfirmModalOpen(false)
+    if (actionType === 'approve' && selectedRequestId) {
+      await approveRequest(selectedRequestId)
+        .unwrap()
+        .then((payload) => console.log('fulfilled', payload))
+        .catch((error) => setShowErrorAlert(error.data.title))
+    } else if (actionType === 'decline' && selectedRequestId) {
+      await declineRequest(selectedRequestId)
+        .unwrap()
+        .then((payload) => console.log('fulfilled', payload))
+        .catch((error) => setShowErrorAlert(error.data.title))
+    }
+    setLoaded(Date.now())
+    setIsLoading(false)
+  }
 
   return (
     <main className="page-main-container">
+      <PageSnackbar
+        open={showErrorAlert !== ''}
+        onCloseNotification={onErrorAlertClose}
+        severity="error"
+        title={t('content.semantichub.alerts.alertErrorTitle')}
+        description={showErrorAlert}
+        showIcon={true}
+        vertical={'bottom'}
+        horizontal={'left'}
+      />
       <CompanyDetailOverlay
         {...{
           openDialog: overlayOpen,
@@ -91,7 +102,10 @@ export default function RegistrationRequests() {
       />
       <ConfirmationOverlay
         openDialog={confirmModalOpen}
-        handleOverlayClose={() => setConfirmModalOpen(false)}
+        handleOverlayClose={() => {
+          setIsLoading(false)
+          setConfirmModalOpen(false)
+        }}
         handleConfirmClick={() => makeActionSelectedRequest()}
       />
 
@@ -112,68 +126,16 @@ export default function RegistrationRequests() {
 
       {/* Table component */}
       <div className={'table-container'}>
-        <Table
-          {...{
-            loading,
-            rows: registrationRequests,
-            columns: columns,
-            rowsCount: registrationRequests.length,
-            toolbarVariant: 'premium',
-            title: t('content.admin.registration-requests.tabletitle'),
-            headerHeight: 76,
-            rowHeight: 100,
-            hideFooter: true,
-            disableColumnFilter: true,
-            disableColumnMenu: true,
-            disableColumnSelector: true,
-            disableDensitySelector: true,
-            disableSelectionOnClick: true,
-            onCellClick: (params: GridCellParams) => onTableCellClick(params),
-            toolbar: {
-              filter: [
-                {
-                  name: 'state',
-                  values: [
-                    {
-                      value: 'pending',
-                      label: t('content.admin.registration-requests.pending')
-                        .trim()
-                        .replace(/^\w/, (c) => c.toUpperCase()),
-                    },
-                    {
-                      value: 'confirmed',
-                      label: t(
-                        'content.admin.registration-requests.cellconfirmed'
-                      )
-                        .trim()
-                        .replace(/^\w/, (c) => c.toUpperCase()),
-                    },
-                    {
-                      value: 'declined',
-                      label: t(
-                        'content.admin.registration-requests.celldeclined'
-                      )
-                        .trim()
-                        .replace(/^\w/, (c) => c.toUpperCase()),
-                    },
-                  ],
-                },
-              ],
-            },
-          }}
-          getRowId={(row) => uniqueId(row.applicationId)}
+        <RequestList
+          fetchHook={useFetchCompanySearchQuery}
+          fetchHookArgs={{ expr }}
+          onSearch={setExpr}
+          onApproveClick={onApproveClick}
+          onDeclineClick={onDeclineClick}
+          isLoading={isLoading}
+          onTableCellClick={onTableCellClick}
+          loaded={loaded}
         />
-      </div>
-      <div className="load-more-button-container">
-        {paginationData.totalElements > pageSize * (currentPage + 1) &&
-          paginationData.totalElements! > pageSize && (
-            <Button
-              size="medium"
-              onClick={() => setCurrentPage((prevState) => prevState + 1)}
-            >
-              {t('content.partnernetwork.loadmore')}
-            </Button>
-          )}
       </div>
     </main>
   )