+++ cx-portal/src/components/pages/Admin/components/RegistrationRequests/registrationTableColumns.tsx
@@ -1,18 +1,26 @@
-import { IconButton, StatusTag, Chip } from 'cx-portal-shared-components'
+import {
+  IconButton,
+  StatusTag,
+  Chip,
+  CircleProgress,
+} from 'cx-portal-shared-components'
 import { GridColDef } from '@mui/x-data-grid'
 import ArrowForwardIcon from '@mui/icons-material/ArrowForward'
 import ArticleOutlinedIcon from '@mui/icons-material/ArticleOutlined'
-import { RegistrationRequestDataGrid } from 'features/admin/registration/types'
 import dayjs from 'dayjs'
 import uniqueId from 'lodash/uniqueId'
+import { useState } from 'react'
+import { ApplicationRequest } from 'features/admin/applicationRequestApiSlice'
 
 // Columns definitions of Registration Request page Data Grid
 export const RegistrationRequestsTableColumns = (
   translationHook: any,
   onApproveClick: (id: string) => void,
-  onDeclineClick: (id: string) => void
+  onDeclineClick: (id: string) => void,
+  isLoading: boolean
 ): Array<GridColDef> => {
   const { t } = translationHook()
+  const [selectedRowId, setSelectedRowId] = useState<string>('')
 
   return [
     {
@@ -21,26 +29,26 @@ export const RegistrationRequestsTableColumns = (
       description: t(
         'content.admin.registration-requests.columns.requestDescription'
       ),
-      flex: 1,
+      flex: 2,
     },
     {
       field: 'dateCreated',
       headerName: t('content.admin.registration-requests.columns.date'),
-      flex: 1,
-      valueGetter: ({ row }: { row: RegistrationRequestDataGrid }) =>
+      flex: 1.5,
+      valueGetter: ({ row }: { row: ApplicationRequest }) =>
         dayjs(row.dateCreated).format('YYYY-MM-DD'),
     },
 
     {
       field: 'companyInfo',
       headerName: t('content.admin.registration-requests.columns.companyinfo'),
-      flex: 2,
+      flex: 2.5,
       sortable: false,
-      renderCell: ({ row }: { row: RegistrationRequestDataGrid }) => (
+      renderCell: ({ row }: { row: ApplicationRequest }) => (
         <div>
-          <p style={{ margin: '3px 0' }}>{row.companyInfo.companyName}</p>
-          <p style={{ margin: '3px 0' }}>{row.companyInfo.email}</p>
-          <span>{row.companyInfo.bpn}</span>
+          <p style={{ margin: '3px 0' }}>{row.companyName}</p>
+          <p style={{ margin: '3px 0' }}>{row.email}</p>
+          <span>{row.bpn}</span>
         </div>
       ),
     },
@@ -50,7 +58,7 @@ export const RegistrationRequestsTableColumns = (
       flex: 2,
       sortable: false,
       cellClassName: 'documents-column--cell',
-      renderCell: ({ row }: { row: RegistrationRequestDataGrid }) => (
+      renderCell: ({ row }: { row: ApplicationRequest }) => (
         <div className="document-cell-container">
           {row.documents.map((contract) => (
             <div
@@ -88,42 +96,69 @@ export const RegistrationRequestsTableColumns = (
       headerName: t('content.admin.registration-requests.columns.state'),
       align: 'center',
       headerAlign: 'center',
-      flex: 1.2,
+      flex: 1,
       sortable: false,
-      renderCell: ({ row }: { row: RegistrationRequestDataGrid }) => {
-        if (row.status === 'SUBMITTED')
+      renderCell: ({ row }: { row: ApplicationRequest }) => {
+        if (row.applicationStatus === 'SUBMITTED')
           return (
             <div className="state-cell-container">
-              <Chip
-                {...{
-                  color: 'secondary',
-                  variant: 'filled',
-                  label: t('content.admin.registration-requests.buttondecline'),
-                  type: 'decline',
-                  onClick: () => onDeclineClick(row.applicationId),
-                  withIcon: true,
-                }}
-              />
+              {selectedRowId === row.applicationId && isLoading ? (
+                <CircleProgress
+                  size={40}
+                  step={1}
+                  interval={0.1}
+                  colorVariant={'primary'}
+                  variant={'indeterminate'}
+                  thickness={8}
+                />
+              ) : (
+                <>
+                  <Chip
+                    {...{
+                      color: 'secondary',
+                      variant: 'filled',
+                      label: t(
+                        'content.admin.registration-requests.buttondecline'
+                      ),
+                      type: 'decline',
+                      onClick: () => {
+                        setSelectedRowId(row.applicationId)
+                        onDeclineClick(row.applicationId)
+                      },
+                      withIcon: true,
+                    }}
+                  />
 
-              <Chip
-                {...{
-                  color: 'secondary',
-                  variant: 'filled',
-                  label: t('content.admin.registration-requests.buttonconfirm'),
-                  type: 'confirm',
-                  onClick: () => onApproveClick(row.applicationId),
-                  withIcon: true,
-                }}
-              />
+                  <Chip
+                    {...{
+                      color: 'secondary',
+                      variant: 'filled',
+                      label: t(
+                        'content.admin.registration-requests.buttonconfirm'
+                      ),
+                      type: 'confirm',
+                      onClick: () => {
+                        setSelectedRowId(row.applicationId)
+                        onApproveClick(row.applicationId)
+                      },
+                      withIcon: true,
+                    }}
+                  />
+                </>
+              )}
             </div>
           )
         else
           return (
             <div className="state-cell-container">
               <StatusTag
-                color={row.status === 'CONFIRMED' ? 'confirmed' : 'declined'}
+                color={
+                  row.applicationStatus === 'CONFIRMED'
+                    ? 'confirmed'
+                    : 'declined'
+                }
                 label={t(
-                  `content.admin.registration-requests.cell${row.status.toLowerCase()}`
+                  `content.admin.registration-requests.cell${row.applicationStatus.toLowerCase()}`
                 )}
               />
             </div>