+++ cx-portal/src/components/pages/NotFound/index.tsx
@@ -1,5 +1,26 @@
+/********************************************************************************
+ * Copyright (c) 2021,2022 BMW Group AG
+ * Copyright (c) 2021,2022 Contributors to the CatenaX (ng) GitHub Organisation.
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ********************************************************************************/
+
 import StageHeader from 'components/shared/frame/StageHeader'
 import { useTranslation } from 'react-i18next'
+import NotFoundContent from './NotFoundContent'
 
 export default function NotFound() {
   const { t } = useTranslation()
@@ -7,13 +28,7 @@ export default function NotFound() {
     <main>
       <StageHeader title={t('pages.notfound')} />
       <section>
-        <p>{t('content.notfound.message')}</p>
-        <ul>
-          <li>{t('content.notfound.reason.nopermission')}</li>
-          <li>{t('content.notfound.reason.notexisting')}</li>
-          <li>{t('content.notfound.reason.notimplemented')}</li>
-        </ul>
-        <p>{t('content.notfound.further')}</p>
+        <NotFoundContent />
       </section>
     </main>
   )