+++ cx-portal/src/features/admin/userApiSlice.ts
@@ -58,6 +58,22 @@ export interface TenantUserDetails extends TenantUser {
   assignedRoles: UserAppRoles[]
 }
 
+export interface CompanyDetails {
+  bpn: string
+  city: string
+  companyId: string
+  countryAlpha2Code: string
+  countryDe: string
+  name: string
+  region: string
+  shortName: string
+  streetAdditional: string
+  streetName: string
+  streetNumber: string
+  taxId: string
+  zipCode: string
+}
+
 export const apiSlice = createApi({
   reducerPath: 'rtk/admin/users',
   baseQuery: fetchBaseQuery({
@@ -98,6 +114,12 @@ export const apiSlice = createApi({
     fetchUserDetails: builder.query<TenantUserDetails, string>({
       query: (id) => `/api/administration/user/owncompany/users/${id}`,
     }),
+    fetchOwnUserDetails: builder.query<TenantUserDetails, void>({
+      query: () => `/api/administration/user/ownUser`,
+    }),
+    fetchOwnCompanyDetails: builder.query<CompanyDetails, void>({
+      query: () => `/api/administration/companydata/ownCompanyDetails`,
+    }),
   }),
 })
 
@@ -106,6 +128,8 @@ export const {
   useFetchUsersQuery,
   useFetchUsersSearchQuery,
   useFetchUserDetailsQuery,
+  useFetchOwnUserDetailsQuery,
   useAddTenantUsersMutation,
   useRemoveTenantUserMutation,
+  useFetchOwnCompanyDetailsQuery,
 } = apiSlice