+++ cx-portal/src/features/apps/apiSlice.ts
@@ -22,6 +22,11 @@ import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
 import i18next from 'i18next'
 import { apiBaseQuery } from 'utils/rtkUtil'
 
+export type ImageType = {
+  src: string
+  alt?: string
+}
+
 export type AppMarketplaceApp = {
   id: string
   title: string
@@ -32,16 +37,19 @@ export type AppMarketplaceApp = {
   price: string
   rating?: number
   link?: string
+  status?: SubscriptionStatus
+  image?: ImageType
 }
 
 export enum SubscriptionStatus {
   ACTIVE = 'ACTIVE',
   PENDING = 'PENDING',
+  INACTIVE = 'INACTIVE',
 }
 
 export type SubscriptionStatusItem = {
   appId: string
-  appSubscriptionStatus: SubscriptionStatus
+  offerSubscriptionStatus: SubscriptionStatus
 }
 
 export type AppDetails = AppMarketplaceApp & {