+++ cx-portal/src/features/notification/types.ts
@@ -12,6 +12,16 @@ export const initialState: ServiceAccountState = {
 }
 
 export enum NotificationType {
+  INFO = 'INFO',
+  ACTION = 'ACTION',
+  WELCOME = 'WELCOME',
+  WELCOME_USE_CASES = 'WELCOME_USE_CASES',
+  WELCOME_SERVICE_PROVIDER = 'WELCOME_SERVICE_PROVIDER',
+  WELCOME_CONNECTOR_REGISTRATION = 'WELCOME_CONNECTOR_REGISTRATION',
+  WELCOME_APP_MARKETPLACE = 'WELCOME_APP_MARKETPLACE',
+  APP_SUBSCRIPTION_REQUEST = 'APP_SUBSCRIPTION_REQUEST',
+  APP_SUBSCRIPTION_ACTIVATION = 'APP_SUBSCRIPTION_ACTIVATION',
+  CONNECTOR_REGISTERED = 'CONNECTOR_REGISTERED',
   Welcome = 'Welcome',
   WelcomeInvite = 'WelcomeInvite',
   WelcomeUser = 'WelcomeUser',
@@ -40,9 +50,10 @@ export interface NotificationContent {
 export type CXNotification = {
   id: string
   created: Date
-  notificationTypeId: NotificationType
+  typeId: NotificationType
   creatorId?: string
   content?: string
   contentParsed?: NotificationContent
-  read?: boolean
+  isRead?: boolean
+  dueDate?: string
 }