+++ cx-portal/src/services/EnvironmentService.test.ts
@@ -40,6 +40,22 @@ const ENV = [
     idp: 'https://centralidp.demo.catena-x.net/auth',
   },
   {
+    name: 'beta',
+    hostname: 'portal.beta.demo.catena-x.net',
+    origin: 'https://portal.beta.demo.catena-x.net',
+    assets: '/assets',
+    api: 'https://portal-backend.beta.demo.catena-x.net',
+    idp: 'https://centralidp.beta.demo.catena-x.net/auth',
+  },
+  {
+    name: 'pre-prod',
+    hostname: 'portal.pre-prod.demo.catena-x.net',
+    origin: 'https://portal.pre-prod.demo.catena-x.net',
+    assets: '/assets',
+    api: 'https://portal-backend.pre-prod.demo.catena-x.net',
+    idp: 'https://centralidp.pre-prod.demo.catena-x.net/auth',
+  },
+  {
     name: 'prod',
     hostname: 'portal.catena-x.net',
     origin: 'https://portal.catena-x.net',